# swagger.yaml
openapi: 3.0.3
info:
  title: Trade-Back API
  description: |
    High-Performance Trading Backend API
    
    This API provides real-time market data streaming, technical indicators, 
    and historical data management for cryptocurrency trading.
    
    ## Features
    - Real-time price streaming via WebSocket (250-450+ msg/sec)
    - Historical data backfilling and management
    - Technical indicators (Enigma with Fibonacci levels)
    - Market watchlist management with auto-subscription
    - High-frequency data updates with binary protocol option
    - Session-based authentication for personalized features
    - Automatic symbol loading from Binance on startup
    
    ## WebSocket Protocol
    The WebSocket endpoint supports both JSON and binary protocols for optimal performance.
    Binary protocol reduces message size by 72% and parsing time by 16x.
    
    ## Market Watch Auto-Subscription
    When a symbol is added to the market watch list:
    1. All WebSocket connections with the same session token auto-subscribe
    2. Real-time prices start streaming immediately
    3. Only subscribed symbols receive price updates (efficient bandwidth usage)
    
    ## Rate Limits
    - REST API: 100 requests per minute per IP
    - WebSocket: No rate limit for subscriptions
    - Historical backfill: 10 requests per second (Binance limits)
  version: 1.0.0
  contact:
    name: Trading Team
    email: support@trade-back.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /symbols:
    get:
      tags: [Market Data]
      summary: Get trading symbols
      description: Retrieve a list of available trading symbols with optional filtering
      operationId: getSymbols
      parameters:
        - name: exchange
          in: query
          description: Filter by exchange
          schema:
            type: string
            example: BINANCE
        - name: type
          in: query
          description: Filter by symbol type
          schema:
            type: string
            enum: [spot, futures]
            example: spot
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbols:
                    type: array
                    items:
                      $ref: '#/components/schemas/Symbol'
                  count:
                    type: integer
                    example: 350
...
  /tradingview/search:
    get:
      tags: [TradingView]
      summary: Search symbols
      description: Search for trading symbols compatible with TradingView
      operationId: searchTradingViewSymbols
      parameters:
        - name: query
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: BTC
        - name: type
          in: query
          description: Symbol type filter
          schema:
            type: string
        - name: exchange
          in: query
          description: Exchange filter
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TradingViewSymbol'
  /tradingview/symbols:
    get:
      tags: [TradingView]
      summary: Get symbol info
      description: Get detailed symbol information for TradingView
      operationId: getTradingViewSymbolInfo
      parameters:
        - name: group
          in: query
          required: true
          description: Symbol or group name
          schema:
            type: string
            example: BTCUSDT
      responses:
        '200':
          description: Symbol information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingViewSymbolInfo'
  /tradingview/history:
    get:
      tags: [TradingView]
      summary: Get price history
      description: Get historical price data in TradingView UDF format
      operationId: getTradingViewHistory
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: resolution
          in: query
          required: true
          description: Chart resolution (1, 5, 15, 30, 60, 240, D, W, M)
          schema:
            type: string
        - name: from
          in: query
          required: true
          description: Unix timestamp
          schema:
            type: integer
        - name: to
          in: query
          required: true
          description: Unix timestamp
          schema:
            type: integer
      responses:
        '200':
          description: Historical data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingViewHistory'
  /tradingview/marks:
    get:
      tags: [TradingView]
      summary: Get chart marks
      description: Get marks for the chart (not implemented)
      operationId: getTradingViewMarks
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: resolution
          in: query
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: integer
        - name: to
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chart marks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Session token for authenticated endpoints
...
  schemas:
...
    Symbol:
      type: object
      properties:
        id:
          type: integer
          example: 1
        symbol:
          type: string
          example: BTCUSDT
        exchange:
          type: string
          example: BINANCE
        full_name:
          type: string
          example: Bitcoin/Tether
        instrument_type:
          type: string
          example: SPOT
        base_currency:
          type: string
          example: BTC
        quote_currency:
          type: string
          example: USDT
        is_active:
          type: boolean
          example: true
        min_price_increment:
          type: number
          format: double
          example: 0.01
        min_quantity_increment:
          type: number
          format: double
          example: 0.00001
...
    TradingViewSymbol:
      type: object
      properties:
        symbol:
          type: string
          example: BTCUSDT
        full_name:
          type: string
          example: BINANCE:BTCUSDT
        description:
          type: string
          example: Bitcoin / Tether
        exchange:
          type: string
          example: BINANCE
        type:
          type: string
          example: crypto
    TradingViewSymbolInfo:
      type: object
      properties:
        symbol:
          type: array
          items:
            type: string
          example: ["BTCUSDT"]
        description:
          type: array
          items:
            type: string
          example: ["Bitcoin / Tether"]
        exchange-listed:
          type: array
          items:
            type: string
          example: ["BINANCE"]
        exchange-traded:
          type: array
          items:
            type: string
          example: ["BINANCE"]
        minmov:
          type: array
          items:
            type: integer
          example: [1]
        pricescale:
          type: array
          items:
            type: integer
          example: [100]
        type:
          type: array
          items:
            type: string
          example: ["crypto"]
        session:
          type: array
          items:
            type: string
          example: ["24x7"]
        has_intraday:
          type: array
          items:
            type: boolean
          example: [true]
        supported_resolutions:
          type: array
          items:
            type: array
            items:
              type: string
        intraday_multipliers:
          type: array
          items:
            type: array
            items:
              type: string
        has_seconds:
          type: array
          items:
            type: boolean
        has_daily:
          type: array
          items:
            type: boolean
        has_weekly_and_monthly:
          type: array
          items:
            type: boolean
    TradingViewHistory:
      type: object
      properties:
        s:
          type: string
          enum: [ok, no_data, error]
          example: ok
        t:
          type: array
          description: Time values (Unix timestamps)
          items:
            type: integer
        o:
          type: array
          description: Open prices
          items:
            type: number
        h:
          type: array
          description: High prices
          items:
            type: number
        l:
          type: array
          description: Low prices
          items:
            type: number
        c:
          type: array
          description: Close prices
          items:
            type: number
        v:
          type: array
          description: Volume values
          items:
            type: number
...